Working Grid Dynamic Refactoring see references for grid.component.html & parent component userpage
