+++ MagicSupliesStore/MSApp/.angular/cache/18.2.5/vite/deps/@angular_material_input.js
@@ -14,8 +14,8 @@ import {
   MatLabel,
   MatPrefix,
   MatSuffix
-} from "./chunk-ANY6UA4N.js";
-import "./chunk-FGX3ACNQ.js";
+} from "./chunk-73ZTCSOC.js";
+import "./chunk-2SUMQTOX.js";
 import {
   ErrorStateMatcher,
   MatCommonModule,
@@ -26,7 +26,7 @@ import {
   coerceNumberProperty,
   getSupportedInputTypes,
   normalizePassiveListenerOptions
-} from "./chunk-4NYHPOPO.js";
+} from "./chunk-HRUU6TGP.js";
 import {
   DOCUMENT
 } from "./chunk-EMLNBNZC.js";
