+++ MagicSupliesStore/MSApp/.angular/cache/18.2.5/vite/deps/@angular_material_progress-spinner.js
@@ -1,6 +1,6 @@
 import {
   MatCommonModule
-} from "./chunk-4NYHPOPO.js";
+} from "./chunk-HRUU6TGP.js";
 import {
   CommonModule,
   NgTemplateOutlet
