+++ MagicSupliesStore/MSApp/.angular/cache/18.2.5/vite/deps/@angular_material_table.js
@@ -7,13 +7,13 @@ import {
   _VIEW_REPEATER_STRATEGY,
   _ViewRepeaterOperation,
   isDataSource
-} from "./chunk-NA4NOVE2.js";
+} from "./chunk-XUTCSZJO.js";
 import {
   Directionality,
   MatCommonModule,
   Platform,
   _isNumberValue
-} from "./chunk-4NYHPOPO.js";
+} from "./chunk-HRUU6TGP.js";
 import {
   DOCUMENT
 } from "./chunk-EMLNBNZC.js";
