+++ MagicSupliesStore/MSApp/.angular/cache/18.2.5/vite/deps/@angular_platform-browser_animations_async.js
@@ -51,7 +51,7 @@ var AsyncAnimationRendererFactory = class _AsyncAnimationRendererFactory {
    * @internal
    */
   loadImpl() {
-    const loadFn = () => this.moduleImpl ?? import("./browser-4EYSULCQ.js").then((m) => m);
+    const loadFn = () => this.moduleImpl ?? import("./browser-DUIMGOJJ.js").then((m) => m);
     let moduleImplPromise;
     if (this.loadingSchedulerFn) {
       moduleImplPromise = this.loadingSchedulerFn(loadFn);
