+++ MagicSupliesStore/MSApp/.angular/cache/18.2.5/vite/deps/chunk-HRUU6TGP.js
@@ -84,32 +84,6 @@ import {
   __spreadValues
 } from "./chunk-3OV72XIM.js";
 
-// node_modules/@angular/cdk/fesm2022/coercion.mjs
-function coerceBooleanProperty(value) {
-  return value != null && `${value}` !== "false";
-}
-function coerceNumberProperty(value, fallbackValue = 0) {
-  if (_isNumberValue(value)) {
-    return Number(value);
-  }
-  return arguments.length === 2 ? fallbackValue : 0;
-}
-function _isNumberValue(value) {
-  return !isNaN(parseFloat(value)) && !isNaN(Number(value));
-}
-function coerceArray(value) {
-  return Array.isArray(value) ? value : [value];
-}
-function coerceCssPixelValue(value) {
-  if (value == null) {
-    return "";
-  }
-  return typeof value === "string" ? value : `${value}px`;
-}
-function coerceElement(elementOrRef) {
-  return elementOrRef instanceof ElementRef ? elementOrRef.nativeElement : elementOrRef;
-}
-
 // node_modules/@angular/cdk/fesm2022/platform.mjs
 var hasV8BreakIterator;
 try {
@@ -335,175 +309,6 @@ function _isTestEnvironment() {
   );
 }
 
-// node_modules/@angular/cdk/fesm2022/bidi.mjs
-var DIR_DOCUMENT = new InjectionToken("cdk-dir-doc", {
-  providedIn: "root",
-  factory: DIR_DOCUMENT_FACTORY
-});
-function DIR_DOCUMENT_FACTORY() {
-  return inject(DOCUMENT);
-}
-var RTL_LOCALE_PATTERN = /^(ar|ckb|dv|he|iw|fa|nqo|ps|sd|ug|ur|yi|.*[-_](Adlm|Arab|Hebr|Nkoo|Rohg|Thaa))(?!.*[-_](Latn|Cyrl)($|-|_))($|-|_)/i;
-function _resolveDirectionality(rawValue) {
-  const value = rawValue?.toLowerCase() || "";
-  if (value === "auto" && typeof navigator !== "undefined" && navigator?.language) {
-    return RTL_LOCALE_PATTERN.test(navigator.language) ? "rtl" : "ltr";
-  }
-  return value === "rtl" ? "rtl" : "ltr";
-}
-var Directionality = class _Directionality {
-  constructor(_document) {
-    this.value = "ltr";
-    this.change = new EventEmitter();
-    if (_document) {
-      const bodyDir = _document.body ? _document.body.dir : null;
-      const htmlDir = _document.documentElement ? _document.documentElement.dir : null;
-      this.value = _resolveDirectionality(bodyDir || htmlDir || "ltr");
-    }
-  }
-  ngOnDestroy() {
-    this.change.complete();
-  }
-  static {
-    this.ɵfac = function Directionality_Factory(__ngFactoryType__) {
-      return new (__ngFactoryType__ || _Directionality)(ɵɵinject(DIR_DOCUMENT, 8));
-    };
-  }
-  static {
-    this.ɵprov = ɵɵdefineInjectable({
-      token: _Directionality,
-      factory: _Directionality.ɵfac,
-      providedIn: "root"
-    });
-  }
-};
-(() => {
-  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(Directionality, [{
-    type: Injectable,
-    args: [{
-      providedIn: "root"
-    }]
-  }], () => [{
-    type: void 0,
-    decorators: [{
-      type: Optional
-    }, {
-      type: Inject,
-      args: [DIR_DOCUMENT]
-    }]
-  }], null);
-})();
-var Dir = class _Dir {
-  constructor() {
-    this._dir = "ltr";
-    this._isInitialized = false;
-    this.change = new EventEmitter();
-  }
-  /** @docs-private */
-  get dir() {
-    return this._dir;
-  }
-  set dir(value) {
-    const previousValue = this._dir;
-    this._dir = _resolveDirectionality(value);
-    this._rawDir = value;
-    if (previousValue !== this._dir && this._isInitialized) {
-      this.change.emit(this._dir);
-    }
-  }
-  /** Current layout direction of the element. */
-  get value() {
-    return this.dir;
-  }
-  /** Initialize once default value has been set. */
-  ngAfterContentInit() {
-    this._isInitialized = true;
-  }
-  ngOnDestroy() {
-    this.change.complete();
-  }
-  static {
-    this.ɵfac = function Dir_Factory(__ngFactoryType__) {
-      return new (__ngFactoryType__ || _Dir)();
-    };
-  }
-  static {
-    this.ɵdir = ɵɵdefineDirective({
-      type: _Dir,
-      selectors: [["", "dir", ""]],
-      hostVars: 1,
-      hostBindings: function Dir_HostBindings(rf, ctx) {
-        if (rf & 2) {
-          ɵɵattribute("dir", ctx._rawDir);
-        }
-      },
-      inputs: {
-        dir: "dir"
-      },
-      outputs: {
-        change: "dirChange"
-      },
-      exportAs: ["dir"],
-      standalone: true,
-      features: [ɵɵProvidersFeature([{
-        provide: Directionality,
-        useExisting: _Dir
-      }])]
-    });
-  }
-};
-(() => {
-  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(Dir, [{
-    type: Directive,
-    args: [{
-      selector: "[dir]",
-      providers: [{
-        provide: Directionality,
-        useExisting: Dir
-      }],
-      host: {
-        "[attr.dir]": "_rawDir"
-      },
-      exportAs: "dir",
-      standalone: true
-    }]
-  }], null, {
-    change: [{
-      type: Output,
-      args: ["dirChange"]
-    }],
-    dir: [{
-      type: Input
-    }]
-  });
-})();
-var BidiModule = class _BidiModule {
-  static {
-    this.ɵfac = function BidiModule_Factory(__ngFactoryType__) {
-      return new (__ngFactoryType__ || _BidiModule)();
-    };
-  }
-  static {
-    this.ɵmod = ɵɵdefineNgModule({
-      type: _BidiModule,
-      imports: [Dir],
-      exports: [Dir]
-    });
-  }
-  static {
-    this.ɵinj = ɵɵdefineInjector({});
-  }
-};
-(() => {
-  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(BidiModule, [{
-    type: NgModule,
-    args: [{
-      imports: [Dir],
-      exports: [Dir]
-    }]
-  }], null, null);
-})();
-
 // node_modules/@angular/cdk/fesm2022/keycodes.mjs
 var ENTER = 13;
 var SHIFT = 16;
@@ -522,6 +327,32 @@ function hasModifierKey(event, ...modifiers) {
     return modifiers.some((modifier) => event[modifier]);
   }
   return event.altKey || event.shiftKey || event.ctrlKey || event.metaKey;
+}
+
+// node_modules/@angular/cdk/fesm2022/coercion.mjs
+function coerceBooleanProperty(value) {
+  return value != null && `${value}` !== "false";
+}
+function coerceNumberProperty(value, fallbackValue = 0) {
+  if (_isNumberValue(value)) {
+    return Number(value);
+  }
+  return arguments.length === 2 ? fallbackValue : 0;
+}
+function _isNumberValue(value) {
+  return !isNaN(parseFloat(value)) && !isNaN(Number(value));
+}
+function coerceArray(value) {
+  return Array.isArray(value) ? value : [value];
+}
+function coerceCssPixelValue(value) {
+  if (value == null) {
+    return "";
+  }
+  return typeof value === "string" ? value : `${value}px`;
+}
+function coerceElement(elementOrRef) {
+  return elementOrRef instanceof ElementRef ? elementOrRef.nativeElement : elementOrRef;
 }
 
 // node_modules/@angular/cdk/fesm2022/observers.mjs
@@ -3138,6 +2969,175 @@ var A11yModule = class _A11yModule {
   }], null);
 })();
 
+// node_modules/@angular/cdk/fesm2022/bidi.mjs
+var DIR_DOCUMENT = new InjectionToken("cdk-dir-doc", {
+  providedIn: "root",
+  factory: DIR_DOCUMENT_FACTORY
+});
+function DIR_DOCUMENT_FACTORY() {
+  return inject(DOCUMENT);
+}
+var RTL_LOCALE_PATTERN = /^(ar|ckb|dv|he|iw|fa|nqo|ps|sd|ug|ur|yi|.*[-_](Adlm|Arab|Hebr|Nkoo|Rohg|Thaa))(?!.*[-_](Latn|Cyrl)($|-|_))($|-|_)/i;
+function _resolveDirectionality(rawValue) {
+  const value = rawValue?.toLowerCase() || "";
+  if (value === "auto" && typeof navigator !== "undefined" && navigator?.language) {
+    return RTL_LOCALE_PATTERN.test(navigator.language) ? "rtl" : "ltr";
+  }
+  return value === "rtl" ? "rtl" : "ltr";
+}
+var Directionality = class _Directionality {
+  constructor(_document) {
+    this.value = "ltr";
+    this.change = new EventEmitter();
+    if (_document) {
+      const bodyDir = _document.body ? _document.body.dir : null;
+      const htmlDir = _document.documentElement ? _document.documentElement.dir : null;
+      this.value = _resolveDirectionality(bodyDir || htmlDir || "ltr");
+    }
+  }
+  ngOnDestroy() {
+    this.change.complete();
+  }
+  static {
+    this.ɵfac = function Directionality_Factory(__ngFactoryType__) {
+      return new (__ngFactoryType__ || _Directionality)(ɵɵinject(DIR_DOCUMENT, 8));
+    };
+  }
+  static {
+    this.ɵprov = ɵɵdefineInjectable({
+      token: _Directionality,
+      factory: _Directionality.ɵfac,
+      providedIn: "root"
+    });
+  }
+};
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(Directionality, [{
+    type: Injectable,
+    args: [{
+      providedIn: "root"
+    }]
+  }], () => [{
+    type: void 0,
+    decorators: [{
+      type: Optional
+    }, {
+      type: Inject,
+      args: [DIR_DOCUMENT]
+    }]
+  }], null);
+})();
+var Dir = class _Dir {
+  constructor() {
+    this._dir = "ltr";
+    this._isInitialized = false;
+    this.change = new EventEmitter();
+  }
+  /** @docs-private */
+  get dir() {
+    return this._dir;
+  }
+  set dir(value) {
+    const previousValue = this._dir;
+    this._dir = _resolveDirectionality(value);
+    this._rawDir = value;
+    if (previousValue !== this._dir && this._isInitialized) {
+      this.change.emit(this._dir);
+    }
+  }
+  /** Current layout direction of the element. */
+  get value() {
+    return this.dir;
+  }
+  /** Initialize once default value has been set. */
+  ngAfterContentInit() {
+    this._isInitialized = true;
+  }
+  ngOnDestroy() {
+    this.change.complete();
+  }
+  static {
+    this.ɵfac = function Dir_Factory(__ngFactoryType__) {
+      return new (__ngFactoryType__ || _Dir)();
+    };
+  }
+  static {
+    this.ɵdir = ɵɵdefineDirective({
+      type: _Dir,
+      selectors: [["", "dir", ""]],
+      hostVars: 1,
+      hostBindings: function Dir_HostBindings(rf, ctx) {
+        if (rf & 2) {
+          ɵɵattribute("dir", ctx._rawDir);
+        }
+      },
+      inputs: {
+        dir: "dir"
+      },
+      outputs: {
+        change: "dirChange"
+      },
+      exportAs: ["dir"],
+      standalone: true,
+      features: [ɵɵProvidersFeature([{
+        provide: Directionality,
+        useExisting: _Dir
+      }])]
+    });
+  }
+};
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(Dir, [{
+    type: Directive,
+    args: [{
+      selector: "[dir]",
+      providers: [{
+        provide: Directionality,
+        useExisting: Dir
+      }],
+      host: {
+        "[attr.dir]": "_rawDir"
+      },
+      exportAs: "dir",
+      standalone: true
+    }]
+  }], null, {
+    change: [{
+      type: Output,
+      args: ["dirChange"]
+    }],
+    dir: [{
+      type: Input
+    }]
+  });
+})();
+var BidiModule = class _BidiModule {
+  static {
+    this.ɵfac = function BidiModule_Factory(__ngFactoryType__) {
+      return new (__ngFactoryType__ || _BidiModule)();
+    };
+  }
+  static {
+    this.ɵmod = ɵɵdefineNgModule({
+      type: _BidiModule,
+      imports: [Dir],
+      exports: [Dir]
+    });
+  }
+  static {
+    this.ɵinj = ɵɵdefineInjector({});
+  }
+};
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(BidiModule, [{
+    type: NgModule,
+    args: [{
+      imports: [Dir],
+      exports: [Dir]
+    }]
+  }], null, null);
+})();
+
 // node_modules/@angular/cdk/fesm2022/cdk.mjs
 var VERSION = new Version("18.2.5");
 
@@ -5146,12 +5146,6 @@ var _MatInternalFormField = class __MatInternalFormField {
 })();
 
 export {
-  coerceBooleanProperty,
-  coerceNumberProperty,
-  _isNumberValue,
-  coerceArray,
-  coerceCssPixelValue,
-  coerceElement,
   Platform,
   getSupportedInputTypes,
   normalizePassiveListenerOptions,
@@ -5161,17 +5155,23 @@ export {
   _getFocusedElementPierceShadowDom,
   _getEventTarget,
   _isTestEnvironment,
-  Directionality,
-  BidiModule,
   ESCAPE,
   hasModifierKey,
+  coerceBooleanProperty,
+  coerceNumberProperty,
+  _isNumberValue,
+  coerceArray,
+  coerceCssPixelValue,
+  coerceElement,
   ObserversModule,
   InteractivityChecker,
   FocusTrapFactory,
   FocusMonitor,
   A11yModule,
+  Directionality,
+  BidiModule,
   MatCommonModule,
   _ErrorStateTracker,
   ErrorStateMatcher
 };
-//# sourceMappingURL=chunk-4NYHPOPO.js.map
+//# sourceMappingURL=chunk-HRUU6TGP.js.map
