+++ MagicSupliesStore/MSApp/src/app/users-page/userspage.component.html
@@ -3,7 +3,9 @@
     <!-- //crear un modelo para "agregar" que abra un modelo para poder agregar un usuario, 
      debe tener un boton para agregar o cancelar y si le das "agregar" use el metodo post 
      para golpear nuestra API// -->
-    <app-gridcomponent #gridComponent (buttonEditSelected)="UpdateUser($event)" (buttonDeleteSelected)="DeleteUser($event)" [dataSourceGridInput]="dataSourceUserPage" ></app-gridcomponent>
+    <app-gridcomponent #gridComponent (buttonEditSelected)="UpdateUser($event)" 
+    (buttonDeleteSelected)="DeleteUser($event)" [dataSourceGridInput]="dataSourceUserPage" 
+    [columns]="userColumns" [headers]="userHeaders"></app-gridcomponent>
     <div class="example-button-container">
       <p-button mat-fab extended (click)="AddOrEditUser()">
         <mat-icon>add</mat-icon>
