+++ MagicSupliesStore/MSApp/src/components/gridcomponent/gridcomponent.html
@@ -1,52 +1,22 @@
-
 <table mat-table [dataSource]="dataSourceGridInput" class="mat-elevation-z8">
 
-    <!--- Note that these columns can be defined in any order.
+  <!--- Note that these columns can be defined in any order.
           The actual rendered columns are set as a property on the row definition" -->
-  
-    <!-- Position Column -->
-    <ng-container matColumnDef="ID">
-      <th mat-header-cell *matHeaderCellDef> No. </th>
-      <td mat-cell *matCellDef="let element"> {{element.id}} </td>
-    </ng-container>
-  
-    <!-- Name Column -->
-    <ng-container matColumnDef="UserName">
-      <th mat-header-cell *matHeaderCellDef> Name </th>
-      <td mat-cell *matCellDef="let element"> {{element.userName}} </td>
-    </ng-container>
-  
-    <!-- Weight Column -->
-    <ng-container matColumnDef="Password">
-      <th mat-header-cell *matHeaderCellDef> Weight </th>
-      <td mat-cell *matCellDef="let element"> {{element.password}} </td>
-    </ng-container>
-  
-    <!-- Symbol Column -->
-    <ng-container matColumnDef="RoleID">
-      <th mat-header-cell *matHeaderCellDef> RoleID </th>
-      <td mat-cell *matCellDef="let element"> {{element.roleID}} </td>
-    </ng-container>
 
-    <ng-container matColumnDef="Mail">
-        <th mat-header-cell *matHeaderCellDef> Mail </th>
-        <td mat-cell *matCellDef="let element"> {{element.mail}} </td>
-      </ng-container>
+  //Reformat of the columns by usinc dynamic values
+  <ng-container *ngFor="let column of columns" [matColumnDef]="column.key">
+    <th mat-header-cell *matHeaderCellDef> {{ column.header }} </th>
+    <td mat-cell *matCellDef="let element"> {{ element[column.key] }} </td>
+  </ng-container>
 
-      <ng-container matColumnDef="CreationDate">
-        <th mat-header-cell *matHeaderCellDef> CreationDate </th>
-        <td mat-cell *matCellDef="let element"> {{element.creationDate}} </td>
-      </ng-container>
-      
-      <ng-container matColumnDef="Actions">
-        <th mat-header-cell *matHeaderCellDef> Actions </th>
-        <td mat-cell *matCellDef="let element"><p-button  (onClick)="SelectEditButtonEvent(element)" >Edit</p-button>
-          &nbsp;<p-button (onClick)="SelectDeleteButtonEvent(element)" >Delete</p-button></td>
+  <ng-container matColumnDef="actions">
+    <th mat-header-cell *matHeaderCellDef> Actions </th>
+    <td mat-cell *matCellDef="let element"><p-button (onClick)="SelectEditButtonEvent(element)">Edit</p-button>
+      &nbsp;<p-button (onClick)="SelectDeleteButtonEvent(element)">Delete</p-button></td>
 
-      </ng-container>
+  </ng-container>
 
-  
-    <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
-    <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
-  </table>
-  
+
+  <tr mat-header-row *matHeaderRowDef="headers"></tr>
+  <tr mat-row *matRowDef="let row; columns: headers;"></tr>
+</table>
